Break up firmware cache and simulator sender
