+++ apps/server/tests/integration/test_refactor_contract_regressions.py
@@ -24,7 +24,8 @@ from vibesensor.infra.runtime.client_snapshot import (
 )
 from vibesensor.infra.runtime.registry import ClientRegistry
 from vibesensor.infra.runtime.registry_updates import _JITTER_EMA_ALPHA, _RESTART_SEQ_GAP
-from vibesensor.use_cases.updates.firmware_cache import FirmwareCacheConfig, GitHubReleaseFetcher
+from vibesensor.use_cases.updates.firmware_release_fetcher import GitHubReleaseFetcher
+from vibesensor.use_cases.updates.firmware_types import FirmwareCacheConfig
 from vibesensor.use_cases.updates.release_fetcher import (
     GitHubAPIClient,
     ReleaseFetcherConfig,
