+++ apps/server/vibesensor/adapters/simulator/__init__.py
@@ -1,1 +1,8 @@
-"""VibeSensor simulator package."""
+"""VibeSensor simulator package.
+
+Module layout:
+- ``sim_client.py``: pure simulated sensor state and frame generation
+- ``sim_scene.py``: pure multi-sensor road-scene mutation logic
+- ``sim_runtime.py``: asyncio UDP protocols and runtime loops
+- ``sim_sender.py``: CLI entry point and top-level orchestration
+"""
